requirements.txt
python-dotenv
flask
flask-SQLAlchemy
flask-CORS
